requirements.txt
fastapi[all]
environ-config
pymongo[srv]
